% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fnr}
\alias{fnr}
\title{False negative rate}
\usage{
fnr(tp, fn)
}
\arguments{
\item{tp}{Number of true-positives}

\item{fn}{Number of false-neagtives}
}
\value{
The value of the true positive rate
}
\description{
This function returns the false negative rate given the number of true-positives and false-negatives.
The respective formula is fnr = fn / (fn + tp)
}
\examples{
#          Reference
# Prediction    1    0
#          1  892  398
#          0  446 2278
tpr(tp = 892, fn = 446)

}
