% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_metrics_curves.R
\name{model_metrics_curves}
\alias{model_metrics_curves}
\title{Plot model metrics across classification threshold}
\usage{
model_metrics_curves(df, metrics, plot_title, threshold = NA)
}
\arguments{
\item{df}{- The tibble that can be generated with the package function [get_threshold_data()]}

\item{metrics}{- A selection of metrics that will be plotted. Note, that some of the returned metrics mean
principally the same: Recall = Sensitivity, Precision = PPV, BalancedAccuracy = roc_auc}

\item{plot_title}{- A title for your plot}

\item{threshold}{- The threshold that should be marked with a vertical dashed line. Optional.}
}
\value{
A ggplot2 object is returned which can be extended by further ggplot functions, be showed and saved.
}
\description{
This function uses the tibble that can be generated with the package function [get_threshold_data()] and plots
the values of selected metrics in dependency of the threshold.
This can be inspiring to evaluate the selection of a threshold for different learning aims.
A selected threshold value can be provided to set a vertical mark in the plot
}
\examples{
y_true <- sample(c(0,1), replace = TRUE, size = 1000)
y_predicted <- runif(1000)
data <- get_threshold_data(truth = y_true, prediction = y_predicted)
plot <- model_metrics_curves(df = data,
                             metrics = c("Sensitivity", "F1", "Balanced Accuracy"),
                             plot_title = "SENS, F1 and balanced Accuracy across classification threshold",
                             threshold = 0.5)
show(plot)

}
