% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_roc_curve.R
\name{pretty_roc_curve}
\alias{pretty_roc_curve}
\title{Plot pretty ROC curve}
\usage{
pretty_roc_curve(
  df,
  plot_title,
  x_col = "fpr",
  y_col = "tpr",
  col_col = "threshold",
  auc_col = "roc_auc",
  colors = c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F", "yellow", "#FF7F00",
    "red", "#7F0000")
)
}
\arguments{
\item{df}{- The tibble that can be generated with the package function [get_threshold_data()]}

\item{plot_title}{- A title for your plot}

\item{x_col}{- The column name of the values 2B plotted on the x-axis. Default = "fpr".}

\item{y_col}{- The column name of the values 2B plotted on the x-axis. Default = "tpr".}

\item{col_col}{- The column name of the values for coloring the points. Default = "threshold".}

\item{auc_col}{- The column name of the values that contain the AUC. Default = "roc-auc". Set to `NA` if you do
not want to show the AUC line}

\item{colors}{- A vector of colors from which a gradient will be generated.
Default: c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F", "yellow", "#FF7F00", "red", "#7F0000")}
}
\value{
A list containing the plot as ggplot2, the maximal AUC (or [auc_col]), the index of this max value in the
        provided [tibble()], the threshold that corresponds to this max value.
}
\description{
This function plots a pretty ROC (receiver-operator curve) using the values from the [tidyverse::tibble()] that
can be generated with the package function [get_threshold_data()].
It colors the curve according to the value of the threshold, and annotates and labels the threshold with maximal AUC.
}
\examples{
y_true <- sample(c(0,1), replace = TRUE, size = 1000)
y_predicted <- runif(1000)
data <- get_threshold_data(truth = y_true, prediction = y_predicted)
roc <- pretty_roc_curve(df = data,
                        plot_title = "ROC curve")
show(roc$plot)

}
