\name{prettyPROC-package}
\alias{prettyPROC-package}
\alias{prettyPROC}
\docType{package}
\title{
\packageTitle{prettyPROC}
}
\description{
\packageDescription{prettyPROC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{prettyPROC}
\packageIndices{prettyPROC}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{prettyPROC}

Maintainer: \packageMaintainer{prettyPROC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
