% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tnr}
\alias{tnr}
\title{True negative rate}
\usage{
tnr(tn, fp)
}
\arguments{
\item{tn}{Number of true-negatives}

\item{fp}{Number of false-positives}
}
\value{
The value of the false positive rate
}
\description{
This function returns the true negative rate given the number of true-negatives and false-positives.
The respective formula is tnr = tn / (tn + fp)
}
\examples{
#          Reference
# Prediction    1    0
#          1  892  398
#          0  446 2278
tpr(tn = 2278, fp = 398)

}
