% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tpr}
\alias{tpr}
\title{True positive rate}
\usage{
tpr(tp, fn)
}
\arguments{
\item{tp}{Number of true-positives}

\item{fn}{Number of false-neagtives}
}
\value{
The value of the true positive rate
}
\description{
This function returns the true positive rate given the number of true-positives and false-negatives.
The respective formula is tpr = tp / (tp + fn)
}
\examples{
#          Reference
# Prediction    1    0
#          1  892  398
#          0  446 2278
tpr(tp = 892, fn = 446)

}
